pyproject.toml

[tool.poetry]
name = "video-analytics-bot"
version = "0.1.0"
description = "Telegram bot for video analytics"
authors = [""]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
aiogram = "3.13.1"
pydantic = "2.9.2"
sqlalchemy = "2.0.35"
psycopg2-binary = "2.9.9"
pydantic-settings = "2.5.2"
asyncpg = "^0.31.0"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"